Work it

+++ src/SnakeServer/GameObjects/Node.cs
@@ -4,8 +4,8 @@ namespace SnakeServer.GameObjects
 {
     public class Node
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         
         public Node(int x, int y)
         {
@@ -13,11 +13,11 @@ namespace SnakeServer.GameObjects
         }
         public Node(double x, double y)
         {
-            X = (int)x; Y = (int)y;
+            X = x; Y = y;
         }
-        public int DistanceTo(Node node)
+        public double DistanceTo(Node node)
         {
-            return (int)(Math.Sqrt(Math.Pow(X - node.X, 2) + Math.Pow(Y - node.Y, 2)));
+            return Math.Sqrt(Math.Pow(X - node.X, 2) + Math.Pow(Y - node.Y, 2));
         }
     }
 }
